requirements.txt
Flask==2.3.2
spacy==3.6.1
transformers==4.33.2
torch==2.0.1
openai==0.28.0
scikit-learn==1.3.0
python-dotenv==1.0.0
